// src/pages/register-worker/types.d.ts
export interface RegisterWorkerFormValues {
  email: string;
  password: string;
  name: string;
  phoneNumber: string;
  alternativePhoneNumber: string;
  dateOfBirth: string;
  gender: string;
  address: string;
  country: string;
  photoRegistration: string;
  graduationCertificate: string;
  qualification: string;
  agencyId: string;
}